WIP ruin everything...: Add Stats, Actions, Statuses to their own components

+++ Assets/Scripts/TurnBased/ITurnBased.cs
@@ -1,6 +1,6 @@
 ﻿public interface ITurnBased
 {
-    float Priority { get; }
+    int Priority { get; }
 
     void OnTurnStart();
     void OnTurnEnd();
